requirements.txt
streamlit
diffusers
torch
torchvision
transformers
opencv-python
pillow
